pyproject.toml
[tool.poetry]
name = "lit-review-pipeline"
version = "0.1.0"
description = "Automated literature review pipeline with harvesting, processing, and screening capabilities"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.0"
requests = "^2.31.0"
streamlit = "^1.28.0"
python-dotenv = "^1.0.0"
scikit-learn = "^1.3.0"
openpyxl = "^3.1.0"
beautifulsoup4 = "^4.12.0"
lxml = "^4.9.0"
langdetect = "^1.0.9"
python-Levenshtein = "^0.20.0"
joblib = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
graphviz = "^0.20.0"
tqdm = "^4.66.0"
xmltodict = "^0.13.0"
python-docx = "^0.8.11"
fuzzywuzzy = "^0.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.284"
pre-commit = "^3.3.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
lit-review = "src.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF"]
ignore = ["S101", "T201", "T203", "PLR0913", "PLR0915", "C901", "PLR0912"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
